code/google_functions/update_sentiment/requirements.txt
tweepy
pandas
google-cloud-bigquery==1.24.0
google-cloud-storage==1.28.0
google-cloud-language==1.3.0
